requirement.txt
streamlit
PyPDF2
python-docx
openai
pandas